requirements.txt
openai>=1.0.0
pytest>=7.0.0
pre-commit>=3.0.0
gitpython>=3.1.0
pyyaml>=6.0
pylint>=2.15.0
bandit>=1.7.4
black>=23.0.0
flake8>=6.0.0
